Extract Position and SimplePosition to module
